+++ Simulation/models/jobs/PredefinedJobGenerator.cpp
@@ -13,7 +13,7 @@ std::unique_ptr<Job> PredefinedJobGenerator::generateJob() {
     int amountPickups = distrEnd(eng);
     std::set<int> stationsToVisit{};
 
-    while (stationsToVisit.size() < amountPickups){
+    while (stationsToVisit.size() < (long unsigned int)amountPickups){
         stationsToVisit.insert(distr(eng));
     }
